Dernières fonctionnalités

+++ front/confirmation.js
@@ -1,48 +1,60 @@
-let contactStorage = JSON.parse(localStorage.getItem("contact"))
-let contact= contactStorage[0]
 let produitStorage = JSON.parse(localStorage.getItem("produit"))
 let products=[];
 
-let totalPrice = JSON.parse(localStorage.getItem("totalPrice"))
-let afficherTotalPrice = document.createElement("p")
-afficherTotalPrice.textContent = "Pour rappel, le prix total de votre commande est de" + totalPrice +"€"
-let essai = document.getElementById("confirm")
-console.log(essai)
-essai.appendChild(afficherTotalPrice)
 
-for(let i = 0; i<produitStorage.length;i++){
-    products.push(produitStorage[i].idProduit)
-}
 
-const aEnvoyer = {
-    products,
+if(products.length ==0){
+    let pageConfirmation = document.getElementById("confirm")
+    pageConfirmation.style.display="none"
     
-    contact
-}
+} else{
+    let contactStorage = JSON.parse(localStorage.getItem("contact"))
+    let contact= contactStorage[0]
 
-console.log(aEnvoyer)
-fetch('http://localhost:3000/api/furniture/order',{
-    method: "POST",
-    body: JSON.stringify(aEnvoyer),
-    headers:{
-        "content-type":"application/JSON"
+    let totalPrice = JSON.parse(localStorage.getItem("totalPrice"))
+    let afficherTotalPrice = document.createElement("p")
+    afficherTotalPrice.textContent = "Pour rappel, le prix total de votre commande est de" + totalPrice +"€"
+    let essai = document.getElementById("confirm")
+    console.log(essai)
+    essai.appendChild(afficherTotalPrice)
+
+    for(let i = 0; i<produitStorage.length;i++){
+        products.push(produitStorage[i].idProduit)
     }
-})
-.then((reponse)=>{
-    const orderData = reponse.json();
-    orderData.then((orderDt) =>{
-        console.log(orderDt)
+
+    const aEnvoyer = {
+        products,
         
-        let orderStorage =[];
-        orderStorage.push(orderDt.orderId)
-        localStorage.setItem("orderId",JSON.stringify(orderStorage))
-        let textOrderId = document.getElementById("orderIdText")
-        textOrderId.textContent = "Votre numéro de commande est le " + orderDt.orderId
-        localStorage.clear()
-        
+        contact
+    }
+    
+    console.log(aEnvoyer)
+    fetch('http://localhost:3000/api/furniture/order',{
+        method: "POST",
+        body: JSON.stringify(aEnvoyer),
+        headers:{
+            "content-type":"application/JSON"
+        }
+    })
+    .then((reponse)=>{
+        const orderData = reponse.json();
+        orderData.then((orderDt) =>{
+            console.log(orderDt)
+            
+            let orderStorage =[];
+            orderStorage.push(orderDt.orderId)
+            localStorage.setItem("orderId",JSON.stringify(orderStorage))
+            let textOrderId = document.getElementById("orderIdText")
+            textOrderId.textContent = "Votre numéro de commande est le " + orderDt.orderId
+            localStorage.clear()
+            
+        })
+    
     })
 
-})
+}
+
+
